Email replier/requirements.txt
flask==2.3.2
openai==0.27.0
python-dotenv==1.0.0
extract-msg==0.5.9
PyPDF2==3.0.0
python-docx==0.8.11
werkzeug==2.3.2
